Update game_manager.js
Made the game 1000x better. If there's already 3 cells in a row, let the cell nudge itself in there. It makes sense conceptually. And it allows for better strategy. It removes the crappy element of the game when your rows fill up and you have to mess up the whole setup because the cell needs to "collide" with something in order for moved to be true. This makes the game infinitely more playable. Trust me.

I love this game. Good work!